requirements.txt
psutil>=5.8.0
pyyaml>=6.0
requests>=2.26.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
python-dateutil>=2.8.2
SpeechRecognition>=3.8.1
opencv-python>=4.5.0
pytest>=7.0.0